f4/usart: UART3/4/5/6 support boilerplate

+++ common/log/log.h
@@ -26,7 +26,7 @@ static inline void _iodev_write(usart_init_t* handle, char *buffer, int size) \
 {                                                                             \
     if (handle)                                                               \
     {                                                                         \
-        PHAL_usartTxBl(handle, (uint16_t *)buffer, size);                     \
+        PHAL_usartTxBl(handle, (uint8_t *)buffer, size);                      \
     }                                                                         \
 }                                                                             \
                                                                               \
